requirements-prod.txt
# Production requirements (includes all from requirements.txt plus production-specific packages)
-r requirements.txt

# Production web server
gunicorn==21.2.0

# AWS deployment
boto3==1.34.34
django-storages==1.14.2

# Environment variables
python-dotenv==1.0.0

# Performance & monitoring
django-redis==5.4.0
redis==5.0.1

# Security
django-cors-headers==4.6.0
